rename file name to expr_with_alias
